A bit more on setup

+++ setup.sh
@@ -1,5 +1,4 @@
 #!/bin/sh
-
 sudo apt-get update
 sudo apt-get install -y libssl-dev zlib1g libpq-dev libcurl4-openssl-dev libcurl4 automake libtool make g++ g++-9 libcrypto++-dev libboost-all-dev libcrypto++-dev libfmt-dev wget cmake libssl-dev libsasl2-dev gdb
 
@@ -9,7 +8,9 @@ git submodule update --init --recursive
 # aegis
 cd submodules/aegis
 sudo ./install-deps.sh
-mkdir build
+if [ ! -d "./build/" ]; then
+  mkdir build
+fi
 cd build
 cmake -DCMAKE_CXX_COMPILER=g++-9 -DCMAKE_CXX_STANDARD=17 ..
 make -j3
@@ -34,23 +35,36 @@ wget http://security.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1-1ubun
 sudo dpkg -i libssl-dev_1.1.1-1ubuntu2.1~18.04.6_amd64.deb libssl1.1_1.1.1-1ubuntu2.1~18.04.6_amd64.deb
 
 # libmongoc
-mkdir mongo
+if [ ! -d "./mongo/" ]; then
+  mkdir mongo
+fi
 cd mongo
-wget https://github.com/mongodb/mongo-c-driver/releases/download/1.17.2/mongo-c-driver-1.17.2.tar.gz
-tar -xzf mongo-c-driver-1.17.2.tar.gz
-rm mongo-c-driver-1.17.2.tar.gz
+if [ ! -d "./mongo-c-driver-1.17.2/" ]; then
+  wget https://github.com/mongodb/mongo-c-driver/releases/download/1.17.2/mongo-c-driver-1.17.2.tar.gz
+  tar -xzf mongo-c-driver-1.17.2.tar.gz
+  rm mongo-c-driver-1.17.2.tar.gz
+else
+  echo "\e[92mlibmongoc directory already exists, not downloading\e[39m"
+fi
 cd mongo-c-driver-1.17.2
-mkdir cmake-build
+
+if [ ! -d "./cmake-build/" ]; then
+  mkdir cmake-build
+fi
 cd cmake-build
-cmake -DENABLE_AUTOMATIC_INIT_AND_CLEANUP=OFF ..
+cmake -DENABLE_AUTOMATIC_INIT_AND_CLEANUP=OFF -DENABLE_TESTS=OFF -DENABLE_EXAMPLES=OFF ..
 make -j4
 sudo cmake --build . --target install
 
 # mongocxx
 cd ../..
-curl -OL https://github.com/mongodb/mongo-cxx-driver/releases/download/r3.6.1/mongo-cxx-driver-r3.6.1.tar.gz
-tar -xzf mongo-cxx-driver-r3.6.1.tar.gz
-rm mongo-cxx-driver-r3.6.1.tar.gz
+if [ ! -d "./mongo-cxx-driver-r3.6.1/" ]; then 
+  curl -OL https://github.com/mongodb/mongo-cxx-driver/releases/download/r3.6.1/mongo-cxx-driver-r3.6.1.tar.gz
+  tar -xzf mongo-cxx-driver-r3.6.1.tar.gz
+  rm mongo-cxx-driver-r3.6.1.tar.gz
+else
+  echo "\e[92mmongocxx directory already exists, not downloading\e[39m"
+fi
 cd mongo-cxx-driver-r3.6.1/build
 
 cmake .. -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=/usr/local -DBSONCXX_POLY_USE_MNMLSTC=1
@@ -58,5 +72,5 @@ sudo cmake --build . --target EP_mnmlstc_core
 make -j4
 sudo cmake --build . --target install
 
-cd ../../../..
-
+cd ../../../../
+./premake.sh
